Rework for improved type roles
* Improve type roles
* Add an improved clone of `Control.Concurrent.STM.TArray`.
