+++ primitive-unlifted.cabal
@@ -23,12 +23,16 @@ library
   exposed-modules:
     Data.Primitive.Unlifted.Class
     Data.Primitive.Unlifted.Array
+    Data.Primitive.Unlifted.Array.ST
     Data.Primitive.Unlifted.Array.Base
+    Data.Primitive.Unlifted.Box
+    Data.Primitive.TArray.Classic
   build-depends:
     , base >=4.14.0.0 && <5
     , bytestring >=0.10.8.2 && <0.11
     , primitive >= 0.7 && <0.8
     , text-short >=0.1.3 && <0.2
+    , array
   hs-source-dirs: src
   ghc-options: -Wall -O2
   default-language: Haskell2010
